requirements/base.txt
# Django & Django Rest Framework
Django==5.2.9
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-simplejwt[crypto]==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
drf-writable-nested==0.7.2
django-import-export==4.3.9

psycopg[binary]==3.2.9






# AWS
boto3==1.40.14

# Deployment
gunicorn==23.0.0

# Miscellaneous
pre-commit==4.3.0

# url builder
yarl==1.23.0